backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.2.3
openpyxl==3.1.5
msoffcrypto-tool==5.4.1
httpx==0.27.2
jinja2==3.1.4
python-dotenv==1.0.1
aiofiles==24.1.0
